feat: add get get data of details hook
